fix dupe module not converting to float64

+++ comparator/duplicatelengthcheck.go
@@ -37,7 +37,7 @@ func (s *DuplicateLengthCheckModule) Run(files ...media.File) (string, string, s
 		return s.Name(), NOCH, "duplicate had insufficient length data, skipping module"
 	}
 	// discard file if new file is shorter than the threshold compared to the duplicate
-	diff := float64(1) - float64(file.RecordedLength / duplicate.RecordedLength)
+	diff := float64(1) - (float64(file.RecordedLength) / float64(duplicate.RecordedLength))
 	if (diff * 100) > float64(settings.Threshold) {
 		return s.Name(), DISC, fmt.Sprintf("new file too short for replacement, (n:%dm/d:%dm) with diff: %d%%, t:%d%%",
 			file.RecordedLength, duplicate.RecordedLength, int64(diff), settings.Threshold)
